break out ExampleComponent into own file
